app/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions==1.23.0
azure-identity==1.23.1
azure-storage-blob==12.26.0
azure-monitor-opentelemetry==1.6.8
aiohttp==3.12.13
# https://github.com/Azure/azure-functions-python-worker/issues/1651
cryptography==43.0.3